Fix bug in page loading:
In the last version, if a element is visible on the screen
when/while the page is loaded/loading, this API will be setted
`data-src` as content for the element `src`, then the browser
will try to show this `src`. The problem is that the `src`,
probably, yet is not loaded, then only a empty space will be
showed.

This update fix this bug. `src` will receive `data-placeholder`
until the `data-src` is loaded. As long as it do not occur, this
API cannot update `src`. While all this occur, the `data-placeholder`
will be showed in the element. Like `data-placeholder` is recommended
most light that `data-src`, the "first loading time" will not be a
problem.