Start library rewrite:
After a long time, the opportunity of use this library
emerged to me. I was not worried, because I already
tested it and I did not make nothing big changes in
it. But, after to run it, I discovered that it is
broke... I do not know the why. I could try fix it,
but I would be wasting an opportunity to improve
it and to practice TypeScript. Because of this,
I decided rewrite this project.